+++ tests/bjson-test.c
@@ -33,6 +33,7 @@
 #include <assert.h>
 #include <errno.h>
 #include <math.h>
+#include <inttypes.h>
 
 #ifdef WIN32
 # include <io.h>
@@ -150,7 +151,7 @@ static int test_bjson_boolean(void *ctx, int value)
   return 1;
 }
 
-static int test_bjson_integer(void *ctx, long long value)
+static int test_bjson_integer(void *ctx, int64_t value)
 {
   if (g_bjson_testMode == TEST_MODE_ENCODE)
   {
@@ -158,7 +159,7 @@ static int test_bjson_integer(void *ctx, long long value)
   }
   else
   {
-    printf("integer: %lld\n", value);
+    printf("integer: %" PRId64 "\n", value);
   }
 
   return 1;
